Time: 0 ms (100.00%) | Memory: 7.4 MB (65.77%) - LeetSync

+++ 476-number-complement/number-complement.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int findComplement(int num) {
-        long long bit=0,k=0;
+        unsigned int bit=0,k=0;
         while(k<=num)bit++,k=pow(2,bit);
         return (k-1)^num;
     }
